"Fixing" TestClick test by removing the subtest for clicking an absolute URL. When the subtest gets fixed, we can add it back.
The absolute URL isn't really testing an absolute url,
and it's breaking the test when Selenium is mounted anywhere else
besides root.
e.g. the test breaks when Selenium is mounted at:
http://localhost:8080/selenium/
but passes when Selenium is mounted at:
http://localhost:8080/

The remaing part of the TestClick file adequetly tests the clicking of a link or button, so this shouldn't be a problem.



r700