[See preceding change for diffs -- I sent the wrong comment by accident.]
Add support for -defaultBrowser commandline argument.  If this argument is used, all following tokens on the command line are used as the browser string for all sessions during this execution of a selenium server.

This is convenient for situations where you want to use a test which has a hardcoded browser string which is not appropriate for your purposes.  For example, SeleneseTestCase always uses *iexplore, but if you are investigating a bug which only occurs under Firefox, you could execute the selenium server with

	-defaultBrowser *firefox

and the test case would execute against Firefox with no need to temporarily modify SeleneseTestCase.



r2548