switchinig to CyberVilliansCA by default with the intent to remove the old genkey.jsp code shortly. The CVCA code is looking pretty solid for FF, Opera, and IE now. The only trick is that instead of using the dangerous-certificate-authority cert, we have to use the cybervilliansCA.cer (checked in here). We will still need to:
1) Document this
2) Provide the ability to supply your own base keystore and CA
3) Make it easy for users to get access to the default CA (perhaps extract it to the working directory the first time SRC runs)

Item #2 will be hard because it means that we need to somehow get the CA in to the FF and Opera profiles. Since Safari and IE use the "global" profile anyway, they aren't as problematic in that they are already painful: we have to install the CA by hand.

r3943