SRC-289, fixing MockPIFrameTest and the build.
The major change here is to change what we do when a Selenium PI frame first loads.  What we were doing was taking the posted "START" and adding it as a command result "OK"; this would confuse the server if the page load "START" was processed before the open "OK" in the previous page, especially since receiving an "OK" automatically marks the page as NOT just loaded.

The current change does two big things:
1) changes the page load logic to more closely match the JS version of the logic.  Pages are marked as loaded when they report a load; they are marked as "not just loaded" when a new command comes in from the client driver, NOT when an "OK" command result comes in from the browser.
2) Throw away the result from browser requests whose GET url contains "seleniumStart=true".  The only command that ever needed a result there was getNewBrowserSession; now he's configured to simply waitForLoad.

With these changes, MockPIFrameTest.testEvilOpen passes on my machine.


r4085