This revision applies Erik Beans' and Jason Huggins' patch, described below and in http://clearspace.openqa.org/docs/DOC-1192
This patch allows the test to specify which of a number of admin-installed profiles on a (remote) machine should be used when starting firefox for that test. It is only integrated with the java client. The RC server gains a command-line flag called -profilesLocation, so that if the machine adminstrator(s) install a profile in $profilesLocation/mytestprofile, the test would call: BrowserConfigurationOptions co = new BrowserConfigurationOptions("profile:mytestprofile"); String url = "http://www.google.com"; selenium = new DefaultSelenium("localhost", RemoteControlConfiguration.getDefaultPort(), "*chrome", url); selenium.start(co); This patch was tested on linux and windows (firefox 2.0), until I wanted to refactor a few things. After that the patch has (so far) only been tested on linux. This patch is a diff file created by svn diff.



r5365