JariBakken: Fix issue caused by FF 3.6 having a lot of stderr output on OS X.
All the output was causing the launch to act funny. Just closing the streams
does the trick for now, but the correct solution might be to read and throw away
the output in a separate thread. Need to look into it more.

This happens only on JRuby - on MRI we're redirecting the streams to /dev/null.

r8199