Fixed an issue whereby a Selenium RC TestSession without an external key could never be cleaned up.
These tests are pretty weak, but illustrate that an exception is not thrown.  I need to figure out how to best mock this to verify behavior.  I suspect this means extracting an interface because I can't figure out how to do a partial mock with the version of EasyMock (2.4) we have locked down.

r13771