KristianRosenvold: A beautiful bugfix
Failure during browser creation on the node in the grid,
would cause updateHubIfNewWebDriverSession to fail with
an exception, which would originally result in the content
length problem from DriverServlet.

The half-created session would later be collected as
timed out, which would further add to the confusion
on the client and the poor bug reporter.

This fixes a bunch of reported bugs, not limited to 3328, 3366
maybe even 2037

r16033