Adding more diagnostics for failed HTTP requests in .NET bindings
This is an attempt to help diagnose the root cause of issue #3719. Please
note that it makes no effort whatsoever to "fix" the issue. This is simply
hoping to provide more information for those who have run into that issue,
and see if all occurrances have the same root cause.