Adding support for native events on Firefox 19.
This includes updated prebuilts for Windows. Linux prebuilts have been
manipulated to not break previous version of Firefox, but new prebuilts
will need to be built.

This updates the Gecko 18 SDK to download 18.0.2, and enables the Gecko 19
SDK downloads.

This renames the webdriver-firefox-10 specific binaries to
webdriver-firefox-esr-previous, and introduces
webdriver-firefox-esr-latest binary shared libraries.