Introducing "browserAttachTimeout" capability to IEDriverServer.exe.
This allows the user to set a capability to control how long the IE driver
will attempt to locate an IE browser window belonging to the iexplore.exe
process launched by the driver.

This is to overcome the bug reported by some users that the IELaunchURL
Windows API reports the incorrect process ID for the launched iexplore.exe
process. In this case, there is nothing the IE driver can do about the
problem as the bug is inside IE itself. However, this will allow a user to
exit the offending situation early, and attempt to recover.

The capability expects an integer value, representing the number of
milliseconds the driver will contininue to look for the newly-launched IE
window. Setting the capability to 0 (which is its default value) will
cause the IE driver to wait indefinitely. The default, "indefinite wait"
behavior aligns with the behavior of the IE driver prior to this commit,
and doesn't fix the issue per se by setting an arbitrary timeout, but
rather gives the user the ability to set their own timeout if desired.

Fixes issue #4541.