Entirely removing the dependency on Castle.Core.dll for .NET support library.
Using the Castle DynamicProxy has been nothing but a thorn in the side of users
of WebDriver.Support.dll. This change implements hard-coded "proxy" objects for
the .NET PageFactory implementation. Hopefully, this will resolve the issues in
the PageFactory for everyone. Nevertheless, there is every possibility that this
will entirely break the PageFactory in some unforseen way for some subset of
users, despite the fact that all of the tests (unit tests, and browser-based
tests) pass. If that is the case, feel free to file issues, if the issues
include code samples that reproduce the problem.

Fixes issue #5221.