Internal code cleanup of IEDriverServer.exe
This commit removes the ATL dependency from IEDriverServer.exe. It does
not remove the dependency on ATL from the internal DLL used by the
executable, only the executable shell itself. It also removes obsolete
APIs from the internally-used IEDriver.dll. Since IEDriverServer.exe is
now the only client of that DLL, and indeed, since the DLL is no longer
produced as a standalone binary, the state-management APIs provided by the
DLL are no longer needed.