Implement ability to specify IE launch API and IE CLI switches
For some cases, it may be necessary to launch IE in Private mode for
testing in parallel on a single node. To support this, there must be a way
to force the IE driver to use the Windows CreateProcess() API to launch IE
so that the user can specify command-line switches. To preserve the
ability to locate the correct process for IE versions 8 and higher, a
registry value must be set, and the IE driver checks for the registry
value before launching IE.

Signed-off-by: Jim Evans <james.h.evans.jr@gmail.com>