Modified to check the interactable state of an element before clicking.
The clear Automation Atom normally checks that an element is interactable
before attempting to clear it. Unfortunately, when the atom is run on a
separate thread in legacy IE, the error states do not get propagated back
to the calling thread properly. To avoid this problem, the driver will
check the interactable state before calling the clear atom. To facilitate
this, a new atom has been exposed for checking the interactable state.