Modified to check the editable state of an element before clearing.
The clear Automation Atom normally checks that an element is interactable
and editable before attempting to clear it. A previous change fixed the
case for checking the interactable state, but not the editable state. The
driver will now check the editable state as well. To facilitate this,
a new atom has been exposed for checking the editable state.