Prevent double-clicks when using WebElement.click().
When using the requireWindowFocus capability, it's possible that
multiple calls to WebElement.click() might happen too quickly, and they
might be interpreted as a double-click. This change inspects the last
time an element was clicked on, and if it is within the system double-
click time, it waits until that time passes before performing the second
click.