Adding IE-specific atoms to eliminate round-trips through IE's JS engine
The findElement and findElements atoms take a JavaScript object as their
first parameter. Until now, that facilitated a round-trip through the IE
JavaScript engine to be able to create the JavaScript object to pass into
the atom function. We are now creating IE-specific JavaScript functions
that wrap those two atoms, and take simple strings as arguments, which
lets the IE driver call the atom directly without having to call a
separate script that simply creates the JavaScript criteria object.