Refactored BrowserFactory class to use member variables.
Previously, most functionality was passed into the methods of the
BrowserFactory, breaking the principle of encapsulation. This change
also moves the ability to force the use of the Windows CreateProcess
API and to pass command-line parameters to the CreateProcess API to
capabilities of the IE driver, rather than arguments passed via the
command line of IEDriverServer.exe.