Refactoring of Proxy setting handling in IE driver
We no longer need to replace the window procedure to accomplish setting
the proxy on a per-process basis, simplifying the process considerably.
This commit also modifies the ProxyManager class to let it build the proxy
setting string instead of prebuilding it in the new session command
handler. Finally, this commit allows the proxy settings to be passed back
to the language bindings via the return from the getSessionCapabilities
command handler.