Introducing ability to clean browser cache before launching IE.
This commit introduces the ie.ensureCleanSession capability, which will
clear the browser cache, history, and cookies before launching IE. When
using this capability, be aware that this clears the cache for all
running instances of Internet Explorer. Using this capability while
attempting to run multiple instances of the IE driver may cause unexpected
behavior. Note this also will cause a performance drop when launching the
browser, as the driver will wait for the cache clearing process to
complete before actually launching IE.