Adding support for ie.ensureCleanSession capability to .NET bindings.
The .NET bindings do this by means of the EnsureCleanSession property on
the InternetExplorerOptions object. When set to true, this property
instructs the IE driver to clear the system cache for all instances of
Internet Explorer, even those already running when the driven instance is
launched, so care should be used when setting this property. The property
defaults to false.