Refactored calculations to the size of the visible view port.
There are circumstances where IE can have no vertical scroll bar. In
these cases, the IE driver would assume there was a vertical scroll
bar, which would make the actual width of the document larger than the
computed width of the window. This, in turn, would make the driver
assume there was a horizontal scroll bar, which would imply that
absolutely-positioned elements at the bottom of the page would not be
seen as clickable.