Take document mode into account when getting size of view port of IE.
When clicking elements in IE, we need to know if the element is in the
view port, and this includes scroll bar calculations. However for sites
that attempt to hide the vertical scroll bar in IE, this may result in
different sizes based on the document rendering mode. Includes tests for
clicking on a link at the very bottom of the view port.