Fixing HTML dialog logic for redirected documents.
Getting the parent window of an HTML dialog's document can fail,
especially if the document changed out from under us before we could get
the window reference. The canonical case for this is a redirect using
JavaScript. Now we will sleep for a shorttime, then retry to obtain the
reference to the window object. Fixes issue #6224.