Refactoring detection of single-text-node child elements in IE.
In IE, a containing text and no background color will not register as a
hit target unless the mouse cursor is directly over the text of the
element. However, in the case of block elements with children that
include text nodes in addition to other elements, if the first text
node does not contain any text, the correct behavior is to use the
center of the block element for mouse movements. This commit corrects
that behavior.