Meta target for building Marionette atoms, marionette:atoms
Marionette uses a custom protocol similar, but not identical, to the
WebDriver wire protocol.  This patch adds a Marionette protocol
function lookup table for the relevant atoms that it uses.

This isn't ideal because it introduces a tight coupling between
Mozilla's repository and Selenium's.  Instead Marionette should be
importing the fragments directly so that no upstream patches are
needed if Marionette decides to change its protocol.

Mozilla bug 936204 has been filed about this.