Updating .NET ChromeOptions to include latest options for Chrome
Also fixes the AddAdditionalCapability method to correctly add additional
options to the chromeOptions object passed to chromedriver.exe, instead of
to the top-level of the capabilities object. Adding new top-level
capabilities is handled by a new overload of AddAdditionalCapability that
contains a boolean argument to indicate that the capability should be
added to the root of the capabilities object. This handles the 80% case
where new "experimental" options are added by the chromedriver team.