Implementing FindsBySequence attribute for .NET PageFactory
Allows the user to mark a field or property in a page object such that
any FindsBy attributes set on the member are used in sequence, as
specified by the Priority property. Note that when not specifying a
Priority value, or specifying multiple FindsBy attributes with the same
Priority value, the behavior is undefined.