Exposing property on .NET PhantomJSDriverService to specify IP address.
When specifying a port on which PhantomJS should listen for WebDriver
commands, the user has the option to also specify an IP address to which
to bind to listen. This commit exposes that ability.

Also exposes the ability to specify a Selenium Grid hub with which this
instance of PhantomJS should register as a node.

Fixes issue #5755.