.NET DriverService improvment to check response on startup
With this commit, a driver using a driver service will now throw a
WebDriverException if it cannot start the driver service, or if the
service executable returns an invalid response. Note that the checks
for the validation of the repsonses is minimal, checking for the
Content-Type header and that the repsonse code is a 200.

Signed-off-by: Jim Evans <james.h.evans.jr@gmail.com>