Adding URL endpoint for switchToParentFrame
This allows consumers of the webdriver-server module to implement the
switchToParentFrame command. Note that this does **not** mean that the
command is automatically implemented in drivers that use this module; it
simply means that the driver will correctly understand the URL endpoint.
As an example, the IE driver should now properly return "command not
implemented" rather than "unknown command".