Handle the case where a proxied element from PageFactory causes a FluentWait to timeout.
It's common for an ExpectedCondition to include a message containing the output
of "element.toString" in the friendly error message. In the case of a proxied
element created by the PageFactory, this will cause a NoSuchElementException to
be thrown, whereas we expect some string or another to be returned.

Handle this case.