Coercing command line argument values to uppercase
Certain command line arguments require values that are uppercase, notably,
the -log-level and -implementation command line switches. If the values of
these arguments are not uppercase, IEDriverServer.exe will fall back to
the default value for the switch. Since Windows command lines are in the
main case-insensitive, it makes sense to coerce these values to uppercase.