JS: Address flakiness with setting copying a Firefox profile
The third-party package used to copy a template Firefox profile would
somtimes fail to copy all of the files in a directory. This would yield
an incomplete profile causing the FirefoxDriver to fail to initialize
properly (up to 10% failrue rate on OS X 10.9.4 with node 0.10.x). This
change removes the problematic dependency infavor of a simple,
synchronous, recursive copy. At some point I will revisit this change to
support fully asynchronous copying.