Removing unused HasBrowserConnection driver extension
This is related to this issue:
https://code.google.com/p/selenium/issues/detail?id=7315

HasBrowserConnection is a driver extension that is only an option for the Android driver. The
Selendroid JSON Wire implementation doesn't support the browser_connection method anymore,
so all code related to this driver extension can be removed from the Ruby bindings.

Here you can verify that the browser_connection method is not available anymore:
https://github.com/selendroid/selendroid/wiki/JSON-Wire-Protocol:-Supported-Methods
https://github.com/selendroid/selendroid/blob/master/selendroid-server/src/main/java/io/selendroid/server/AndroidServlet.java
https://github.com/selendroid/selendroid/blob/master/selendroid-standalone/src/main/java/io/selendroid/server/SelendroidServlet.java