Fix return types Actions and TouchActions classes.
The challenge of the builder pattern when extending the class
is that the subclass must either redefine all the methods
of the parent class or the subclass will return instances
with a static type signature of the parent class.

For example the following would fail to compile without this
patch: new TouchActions(driver).click().singleTap(element).
Selfishly I would like to see this patch get accepted to
fix these same bugs in my own subclass of the Actions class
that might eventually become its own pull request.

I have filled in the CLA agreement.