Make uploading files via python remote webdriver disabled by default.
Uploading files via remote webdriver was originally added to java
webdriver in the following commit:
https://github.com/SeleniumHQ/selenium/commit/90d9a761df576d8f733628d519e429421dc3a3c7

By default this feature is disabled in java and it can be enabled by
setFileDetector.

Python port of code was done in the commit:
https://github.com/SeleniumHQ/selenium/commit/82f3da579826b5a694395d1e5efbe7bd95c5a2a7
It has ported java feature without possibility of turning it on/off.

This commit ports the rest of the java code, by making files upload
disabled by default and adding possibility of turning it on via setting
FileDetector.