IE: Introduce capability to disable check of mime type on setting cookie
When setting cookies, there is a check in the IE driver to validate
that the page in the browser is, in fact, an HTML page. Despite the fact
that omitting this check can cause unrecoverable crashes in the driver,
there is demand for a mechanism to disable this check for older, legacy
versions of Internet Explorer.

This commit introduces a new capability, "ie.validateCookieDocumentType"
which, when set to false, skips this check when adding a cookie. The
default for this capability when unspecified for a session is true;
users setting this capability to false are considered to be on their own
when encountering issues setting cookies, including, but not limited to,
crashes in the driver executable.

Furthermore, this commit adds a type safe option in the .NET
InternetExplorerOptions class to set this capability. Other language
bindings will need updates from their maintainers to provide these type
safe options.

Fixes issue #1227.