Update crazy-fun to handle Closure's new goog.module syntax in js builds.
We were using Closure's calcdeps.py script to build a dependency graph,
but this has been deprecated and will not be updated for the goog.module
primitive. Rather than fork it, I moved its logic (which is relatively
straight forward) into crazy_fun. Also using this opportunity to merge
the js_deps and js_library rules, which served the same purpose.