Allow to pass firefox_profile as a path.
This commit allows to use a path as the firefox_profile argument
in the Firefox's webdriver.

Currently a FirefoxProfile instance can only be used by a
single webdriver. If one FirefoxProfile instance is used by more
than one webdriver, then temporal directories are removed when
the first webdriver instance quits, raising errors in the rest.

By creating the FirefoxProfile instance inside the WebDriver class we
guarantee that the mapping FirefoxProfile <-> WebDriver is one
to one.