Updated session intialization to more robustly handle malformed JSON.
Previously, in the IE driver, malformed JSON requests when requesting a
new session would cause the driver to crash. The new behavior is that
if the JSON payload for the new session is malformed by not including
a desiredCapabilities object will now throw an exception. If the user
passes a capability value with the wrong data type, it will be logged
by the driver at a warning level, and the default value for the specified
capability will be used.