Adding option to not delete anonymous Firefox profile in .NET
This change adds an option to the .NET FirefoxProfile class so that the
driver will not delete the anonymous profile created by the driver. Since
the driver cannot and should not use an existing profile in situ because
of the multiple instance automation case, this change means that
modifications applied to the anonymous profile can be retained and
used in future anonymous profiles.The implication is that the user can
now make modifications to a profile, and retain those profile modifications
(e.g., cookies) into other future profiles, simulating persistent changes
over multiple browser launches. Fixes issue #7374.