Adding initial implementation of .NET bindings OperaDriver
The .NET bindings will now support the Chromium-based Opera driver without
requiring the use of the Java remote WebDriver server. This driver will
work with Opera 26 and above, and requires the download of the Opera
driver executable. Code cleanup and refactoring will take place under a
separate commit.

Note that there is still no support in the .NET bindings for the
Presto-based Opera without using the remote server, nor is there likely to
be.