Fixing finding elements by XPath in IE9+ with document mode >= IE9
The previous code checked the browser version to ensure that XML
namespaces were implemented for finding by XPath in IE 9 or higher.
However, this would still fail in IE9 if the document mode was below the
document mode for IE9. This commit changes the check to look at document
mode rather than browser version.